Ignore prodtype when building Benchmark and include only rules that are part of a Profile
This is the first step to get rid of prodtype. We introduce a way
to ignore current value during the build phase while being able to
work with existing prodtypes to build/modify profiles.

Also, now the build system won't put a rule into DS if it is not
present in a profile (hidden or real).

Hidden profile is a profile with 'hidden: true' attribute. It won't be
rendered as an XCCDF Profile, but it would bring all its rules into
the DS.

Converter script: convert_prodtype_into_default_profile_entry.py
The remover: convert_prodtype_remove_prodtype.sh